refactor: reorganização da estrutura de models e limpeza de código
- Move todos os models das subpastas para a pasta principal (model/)
- Remove pastas desnecessárias para simplificar a navegação
- Substitui @Autowired por constructor injection no TransacaoController
- Remove código comentado desnecessário
- Atualiza imports em todos os arquivos afetados
- Mantém todas as funcionalidades existentes
