+++ implementacao/moeda/src/main/java/com/example/moeda/model/Professor.java
@@ -1,28 +1,15 @@
-package com.example.moeda.model.professor;
+package com.example.moeda.model;
 
-import com.example.moeda.model.pessoa.Pessoa;
-import com.example.moeda.model.departamento.Departamento;
 import jakarta.persistence.*;
 
 @Entity
 @Table(name = "professores")
 public class Professor extends Pessoa {
-   // @Column(nullable = false, unique = true)
-   // private String matricula;
-
     @ManyToOne
     @JoinColumn(name = "departamento_id", nullable = false)
     private Departamento departamento;
 
     // Getters e Setters
-   // public String getMatricula() {
-   //     return matricula;
-   // }
-
-    //public void setMatricula(String matricula) {
-     //   this.matricula = matricula;
-    //}
-
     public Departamento getDepartamento() {
         return departamento;
     }
@@ -30,4 +17,4 @@ public class Professor extends Pessoa {
     public void setDepartamento(Departamento departamento) {
         this.departamento = departamento;
     }
-}
+}
