+++ implementacao/moeda/src/main/java/com/example/moeda/repository/CursoRepository.java
@@ -1,6 +1,6 @@
 package com.example.moeda.repository;
 
-import com.example.moeda.model.curso.Curso;
+import com.example.moeda.model.Curso;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
