+++ implementacao/moeda/src/main/java/com/example/moeda/repository/EmpresaRepository.java
@@ -1,6 +1,6 @@
 package com.example.moeda.repository;
 
-import com.example.moeda.model.empresa.Empresa;
+import com.example.moeda.model.Empresa;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
