+++ implementacao/moeda/src/main/java/com/example/moeda/repository/InstituicaoRepository.java
@@ -1,6 +1,6 @@
 package com.example.moeda.repository;
 
-import com.example.moeda.model.instituicao.Instituicao;
+import com.example.moeda.model.Instituicao;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
