+++ implementacao/moeda/src/main/java/com/example/moeda/repository/VantagemRepository.java
@@ -1,6 +1,6 @@
 package com.example.moeda.repository;
 
-import com.example.moeda.model.vantagem.Vantagem;
+import com.example.moeda.model.Vantagem;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
